+++ tamr_client/instance.py
@@ -1,4 +1,4 @@
-from tamr_client._types import Instance
+from tamr_client._types import Instance, Session
 
 
 def origin(instance: Instance) -> str:
@@ -10,3 +10,19 @@ def origin(instance: Instance) -> str:
         return f"{instance.protocol}://{instance.host}"
     else:
         return f"{instance.protocol}://{instance.host}:{instance.port}"
+
+
+def version(session: Session, instance: Instance) -> str:
+    """Return the Tamr version for an instance.
+
+    Args:
+        session: Tamr Session
+        instance: Tamr instance
+
+    Returns: Version
+
+    """
+    # Version endpoints are not themselves versioned by design, but they are stable so they are ok to use here.
+    return session.get(f"{origin(instance)}/api/versioned/service/version").json()[
+        "version"
+    ]